+++ tests/corpus/csharp_using_alias/src/MyApp/Models/User.cs
@@ -1,0 +1,16 @@
+namespace MyApp.Models;
+
+public class User
+{
+    public string Name { get; }
+
+    public User(string name)
+    {
+        Name = name;
+    }
+
+    public static string Describe(User user)
+    {
+        return "user " + user.Name;
+    }
+}